requirements.txt
uvicorn
fastapi
libgpiod
RPi.GPIO